Refactor: Migrate to unified payment and QR code models
