+++ app/api/v1/endpoints/admin/alerts.py
@@ -6,7 +6,8 @@ from app.models.admin import Admin
 from app.schemas.admin import AdminFullResponse, AdminProfileUpdate, AdminQRCodeCreate
 # from app.services.logging_service import AuditLoggingService  # Больше не используется
 from app.services.two_phase_commit_service import two_phase_commit_service
-from app.models.payment import PaymentRequest, TransactionStatus    
+from app.models.unified import UnifiedPayment as PaymentRequest
+from app.models.enums import TransactionStatus    
 from app.core.dependencies import get_current_admin_user
 
 from datetime import datetime, timedelta, timezone
