+++ app/services/enhanced_validation_service.py
@@ -6,7 +6,8 @@ from enum import Enum
 from datetime import datetime, timedelta
 from decimal import Decimal, InvalidOperation
 from app.services.error_handling_service import StandardError, ErrorCode, ErrorSeverity
-from app.models.payment import Bank, PaymentRequest
+from app.models.unified import UnifiedPayment as PaymentRequest
+from app.models.payment import Bank
 import logging
 
 logger = logging.getLogger(__name__)
