+++ app/services/two_phase_commit_service.py
@@ -12,7 +12,9 @@ import asyncio
 import aiohttp
 from contextlib import asynccontextmanager
 
-from app.models.payment import PaymentRequest, TransactionStatus, TwoPhaseOperation, Bank
+from app.models.unified import UnifiedPayment as PaymentRequest
+from app.models.enums import TransactionStatus
+from app.models.payment import TwoPhaseOperation, Bank
 from app.schemas.bank import (
     TransactionPrepareRequest, TransactionPrepareResponse,
     TransactionCommitRequest, TransactionCommitResponse,
